# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: zboard API
  version: v0.0.1
  description: |
    zboard single-binary operations console API for subscriber and operator workflows.
    Business APIs use /api/v1. Zero's standard event webhook is exposed at /api/zero/events.
servers:
  - url: http://localhost:8080
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/subscription-rule-sets:
    get:
      tags: [subscriptions, admin]
      summary: List reusable subscription rule sets
      description: Returns a bounded metadata page. Templates reference these records by ID; inline remote entries remain available as a template-only shortcut.
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 100 }, description: Search name, tag, description or URL. }
        - { name: renderer, in: query, schema: { type: string, enum: [znet-sink, clash, sing-box] } }
        - { name: active, in: query, schema: { type: boolean } }
        - { name: id, in: query, schema: { type: integer, format: int64, minimum: 1 }, description: Resolve one selected item. }
        - { name: ids, in: query, schema: { type: string }, description: Resolve up to 64 comma-separated selected IDs. }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Reusable rule-set page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data: { $ref: "#/components/schemas/SubscriptionRuleSetPage" }
    post:
      tags: [subscriptions, admin]
      summary: Create a reusable remote subscription rule set
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionRuleSetWriteRequest" }
      responses:
        "200": { description: Rule set created }
        "400": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    NodeConnectorBearer:
      type: http
      scheme: bearer
      bearerFormat: ZeroNodeAPIKey
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true
        error:
          $ref: "#/components/schemas/ApiErrorDetail"
        timestamp:
          type: string
    ApiErrorDetail:
      type: object
      description: Versioned machine-readable detail included on HTTP error responses. The top-level message remains for backward compatibility.
      required: [version, code]
      properties:
        version:
          type: integer
          enum: [1]
        code:
          type: string
          example: validation_failed
        fields:
          type: object
          description: Field-path to localized actionable message. Present when code is validation_failed.
          additionalProperties:
            type: string
    PageMetadata:
      type: object
      required: [offset, limit, total, next_cursor, previous_cursor]
      properties:
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, format: int64, minimum: 0 }
        next_cursor: { type: string, nullable: true }
        previous_cursor: { type: string, nullable: true }
    PageEnvelope:
      type: object
      required: [page, aggregates, facets]
      properties:
        page: { $ref: "#/components/schemas/PageMetadata" }
        aggregates: { type: object, additionalProperties: true }
        facets: { type: object, additionalProperties: true }
        total: { type: integer, format: int64, minimum: 0, deprecated: true, description: Compatibility alias for page.total. }
        offset: { type: integer, minimum: 0, deprecated: true, description: Compatibility alias for page.offset. }
        limit: { type: integer, minimum: 1, maximum: 200, deprecated: true, description: Compatibility alias for page.limit. }
...
    SubscriptionRuleSet:
      type: object
      required: [id, name, renderer, tag, url, format, interval, is_active, revision, usage_count, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, maxLength: 80 }
        description: { type: string, maxLength: 255 }
        renderer: { type: string, enum: [znet-sink, clash, sing-box] }
        tag: { type: string, minLength: 1, maxLength: 64, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$" }
        url: { type: string, format: uri, maxLength: 2048 }
        behavior: { type: string, enum: [domain, ipcidr, classical], description: Clash-only provider behavior. }
        format: { type: string, enum: [domain_list, cidr_list, zero_rule_ir, zrs, yaml, text, mrs, source, binary] }
        interval: { type: integer, minimum: 60, maximum: 604800 }
        is_active: { type: boolean }
        revision: { type: integer, format: int64, minimum: 1 }
        usage_count: { type: integer, format: int64, minimum: 0, description: Number of templates referencing this rule set. }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SubscriptionRuleSetPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/SubscriptionRuleSet" }
    SubscriptionRuleSetWriteRequest:
      type: object
      additionalProperties: false
      required: [name, renderer, tag, url, interval, is_active]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 255 }
        renderer: { type: string, enum: [znet-sink, clash, sing-box] }
        tag: { type: string, minLength: 1, maxLength: 64, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$" }
        url: { type: string, format: uri, maxLength: 2048 }
        behavior: { type: string, enum: [domain, ipcidr, classical] }
        format: { type: string, enum: [domain_list, cidr_list, zero_rule_ir, zrs, yaml, text, mrs, source, binary] }
        interval: { type: integer, minimum: 60, maximum: 604800 }
        is_active: { type: boolean }
        expected_revision: { type: integer, format: int64, minimum: 1 }
...